Remove nested structure for fixture files
There seems to be a Rails issue where nested folders for fixture files don't seem to work.
The issue seems to be about path sanitization. I may open an issue for this later.

But, for now, I'm removing them. A flat structure seems to work, and lets me use fixture files still. Cool?